01-data-modeling-etl/requirements.txt
# Основные библиотеки
pandas
sqlalchemy
psycopg2-binary
python-dotenv

# Линтеры и Типизация
pylint
mypy
pandas-stubs
sqlalchemy-stubs